HRTIM refactor
