+++ src/pwm/hrtim.rs
@@ -1,25 +1,31 @@
+mod external_event;
+mod fault;
+mod calibration;
+
 use core::marker::PhantomData;
 use core::mem::MaybeUninit;
 
 use fugit::HertzU64;
 
-use crate::comparator::{COMP1, COMP2, COMP3, COMP4, COMP5, COMP6};
-use crate::gpio::gpioa::{PA10, PA11, PA12, PA15, PA8, PA9};
-use crate::gpio::gpiob::{PB0, PB10, PB11, PB12, PB13, PB14, PB15};
-use crate::gpio::gpioc::{PC10, PC6, PC7, PC8, PC9};
-use crate::gpio::{self, AF3};
-use crate::gpio::{Alternate, AF13};
+use crate::gpio::gpioa::{PA10, PA11, PA8, PA9};
+use crate::gpio::gpiob::{PB12, PB13, PB14, PB15};
+use crate::gpio::gpioc::{PC6, PC7, PC8, PC9};
+use crate::gpio::{Alternate, AF13, AF3};
+use crate::pwm::hrtim::fault::{FaultAction, FaultSource};
 use crate::stm32::{
     HRTIM_COMMON, HRTIM_MASTER, HRTIM_TIMA, HRTIM_TIMB, HRTIM_TIMC, HRTIM_TIMD, HRTIM_TIME,
     HRTIM_TIMF,
 };
 
-use super::{
-    ActiveHigh, Alignment, ComplementaryImpossible, FaultMonitor, Pins, Polarity, Pwm,
-    PwmPinEnable, TimerType,
+use self::calibration::HrTimOngoingCalibration;
+use self::fault::{
+    FltMonitor1, FltMonitor2, FltMonitor3, FltMonitor4, FltMonitor5, FltMonitor6, FltMonitorSys,
 };
-use crate::rcc::{Enable, GetBusFreq, Rcc, Reset};
-use crate::stm32::RCC;
+
+use super::{
+    ActiveHigh, Alignment, ComplementaryImpossible, Pins, Polarity, Pwm, PwmPinEnable, TimerType,
+};
+use crate::rcc::{GetBusFreq, Rcc};
 use crate::time::Hertz;
 
 pub struct CH1<PSCL>(PhantomData<PSCL>);
@@ -419,7 +425,7 @@ macro_rules! impl_into_es {
     ($dst:ident, $t:ty, $ES:ident) => {
         impl<PSCL> Into<EventSource<PSCL, $dst>> for &$t {
             fn into(self) -> EventSource<PSCL, $dst> {
-                EventSource::$ES(PhantomData)
+                EventSource::$ES{ _x: PhantomData }
             }
         }
     };
@@ -477,7 +483,9 @@ macro_rules! impl_into_neighbor_es {
     ($dst:ident, $src:ident, $cr:ident, $TimEventX:ident) => {
         impl<PSCL> Into<EventSource<PSCL, $dst>> for &$cr<$src, PSCL> {
             fn into(self) -> EventSource<PSCL, $dst> {
-                EventSource::NeighborTimer(NeighborTimerEventSource::$TimEventX(PhantomData))
+                EventSource::NeighborTimer {
+                    n: NeighborTimerEventSource::$TimEventX { _x: PhantomData },
+                }
             }
         }
     };
@@ -569,36 +577,38 @@ pub enum State {
 
 pub enum EventSource<PSCL, DST> {
     /// Compare match with compare register 1 of this timer
-    Cr1(PhantomData<(PSCL, DST)>),
+    Cr1 { _x: PhantomData<(PSCL, DST)> },
 
     /// Compare match with compare register 2 of this timer
-    Cr2(PhantomData<(PSCL, DST)>),
+    Cr2 { _x: PhantomData<(PSCL, DST)> },
 
     /// Compare match with compare register 3 of this timer
-    Cr3(PhantomData<(PSCL, DST)>),
+    Cr3 { _x: PhantomData<(PSCL, DST)> },
 
     /// Compare match with compare register 4 of this timer
-    Cr4(PhantomData<(PSCL, DST)>),
+    Cr4 { _x: PhantomData<(PSCL, DST)> },
 
     /// On complete period
-    Period(PhantomData<(PSCL, DST)>),
+    Period { _x: PhantomData<(PSCL, DST)> },
 
     /// Compare match with compare register 1 of master timer
-    MasterCr1(PhantomData<(PSCL, DST)>),
+    MasterCr1 { _x: PhantomData<(PSCL, DST)> },
 
     /// Compare match with compare register 2 of master timer
-    MasterCr2(PhantomData<(PSCL, DST)>),
+    MasterCr2 { _x: PhantomData<(PSCL, DST)> },
 
     /// Compare match with compare register 3 of master timer
-    MasterCr3(PhantomData<(PSCL, DST)>),
+    MasterCr3 { _x: PhantomData<(PSCL, DST)> },
 
     /// Compare match with compare register 4 of master timer
-    MasterCr4(PhantomData<(PSCL, DST)>),
+    MasterCr4 { _x: PhantomData<(PSCL, DST)> },
 
     /// On complete master period
-    MasterPeriod(PhantomData<(PSCL, DST)>),
+    MasterPeriod { _x: PhantomData<(PSCL, DST)> },
 
-    NeighborTimer(NeighborTimerEventSource<PSCL, DST>),
+    NeighborTimer {
+        n: NeighborTimerEventSource<PSCL, DST>,
+    },
 }
 
 /// Compare events from neighbor timers
@@ -616,7 +626,9 @@ pub enum NeighborTimerEventSource<PSCL, DST> {
     /// |TimD | A CR1  |
     /// |TimE | A CR4  |
     /// |TimF | A CR3  |
-    TimEvent1(PhantomData<(PSCL, DST)>),
+    TimEvent1 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
 
     /// Timer event x
     ///
@@ -642,7 +654,9 @@ pub enum NeighborTimerEventSource<PSCL, DST> {
     /// |TimD | A CR4  |
     /// |TimE | B CR3  |
     /// |TimF | B CR1  |
-    TimEvent2(PhantomData<(PSCL, DST)>),
+    TimEvent2 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
 
     /// Timer event 3
     ///
@@ -655,7 +669,9 @@ pub enum NeighborTimerEventSource<PSCL, DST> {
     /// |TimD | B CR2  |
     /// |TimE | B CR4  |
     /// |TimF | B CR4  |
-    TimEvent3(PhantomData<(PSCL, DST)>),
+    TimEvent3 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
 
     /// Timer event 4
     ///
@@ -668,14 +684,26 @@ pub enum NeighborTimerEventSource<PSCL, DST> {
     /// |TimD | B CR4  |
     /// |TimE | C CR1  |
     /// |TimF | C CR1  |
-    TimEvent4(PhantomData<(PSCL, DST)>),
+    TimEvent4 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
 
     // TODO: Document those
-    TimEvent5(PhantomData<(PSCL, DST)>),
-    TimEvent6(PhantomData<(PSCL, DST)>),
-    TimEvent7(PhantomData<(PSCL, DST)>),
-    TimEvent8(PhantomData<(PSCL, DST)>),
-    TimEvent9(PhantomData<(PSCL, DST)>),
+    TimEvent5 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
+    TimEvent6 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
+    TimEvent7 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
+    TimEvent8 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
+    TimEvent9 {
+        _x: PhantomData<(PSCL, DST)>,
+    },
 }
 
 macro_rules! hr_timer_reset_event_source_common {
@@ -1149,6 +1177,12 @@ macro_rules! hrtim_finalize_body {
                     .sdtrx().bit(deadtime_rising_sign)
                     .dtfx().bits(deadtime_falling_value)
                     .sdtfx().bit(deadtime_falling_sign)
+
+                    // Lock configuration
+                    .dtflkx().set_bit()
+                    .dtfslkx().set_bit()
+                    .dtrlkx().set_bit()
+                    .dtrslkx().set_bit()
                 );
                 tim.$outXr.modify(|_r, w| w.dten().set_bit());
             }
@@ -1629,44 +1663,44 @@ macro_rules! hrtim_out_common {
         let tim = unsafe { &*$TIMX::ptr() };
 
         match $set_event {
-            EventSource::Cr1(..) => tim.$register.modify(|_r, w| w.cmp1().$action()),
-            EventSource::Cr2(..) => tim.$register.modify(|_r, w| w.cmp2().$action()),
-            EventSource::Cr3(..) => tim.$register.modify(|_r, w| w.cmp3().$action()),
-            EventSource::Cr4(..) => tim.$register.modify(|_r, w| w.cmp4().$action()),
-            EventSource::Period(..) => tim.$register.modify(|_r, w| w.per().$action()),
+            EventSource::Cr1 { .. } => tim.$register.modify(|_r, w| w.cmp1().$action()),
+            EventSource::Cr2 { .. } => tim.$register.modify(|_r, w| w.cmp2().$action()),
+            EventSource::Cr3 { .. } => tim.$register.modify(|_r, w| w.cmp3().$action()),
+            EventSource::Cr4 { .. } => tim.$register.modify(|_r, w| w.cmp4().$action()),
+            EventSource::Period { .. } => tim.$register.modify(|_r, w| w.per().$action()),
 
-            EventSource::MasterCr1(..) => tim.$register.modify(|_r, w| w.mstcmp1().$action()),
-            EventSource::MasterCr2(..) => tim.$register.modify(|_r, w| w.mstcmp2().$action()),
-            EventSource::MasterCr3(..) => tim.$register.modify(|_r, w| w.mstcmp3().$action()),
-            EventSource::MasterCr4(..) => tim.$register.modify(|_r, w| w.mstcmp4().$action()),
-            EventSource::MasterPeriod(..) => tim.$register.modify(|_r, w| w.mstper().$action()),
+            EventSource::MasterCr1 { .. } => tim.$register.modify(|_r, w| w.mstcmp1().$action()),
+            EventSource::MasterCr2 { .. } => tim.$register.modify(|_r, w| w.mstcmp2().$action()),
+            EventSource::MasterCr3 { .. } => tim.$register.modify(|_r, w| w.mstcmp3().$action()),
+            EventSource::MasterCr4 { .. } => tim.$register.modify(|_r, w| w.mstcmp4().$action()),
+            EventSource::MasterPeriod { .. } => tim.$register.modify(|_r, w| w.mstper().$action()),
 
-            EventSource::NeighborTimer(es) => match es {
-                NeighborTimerEventSource::TimEvent1(..) => {
+            EventSource::NeighborTimer { n } => match n {
+                NeighborTimerEventSource::TimEvent1 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt1().$action())
                 }
-                NeighborTimerEventSource::TimEvent2(..) => {
+                NeighborTimerEventSource::TimEvent2 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt2().$action())
                 }
-                NeighborTimerEventSource::TimEvent3(..) => {
+                NeighborTimerEventSource::TimEvent3 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt3().$action())
                 }
-                NeighborTimerEventSource::TimEvent4(..) => {
+                NeighborTimerEventSource::TimEvent4 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt4().$action())
                 }
-                NeighborTimerEventSource::TimEvent5(..) => {
+                NeighborTimerEventSource::TimEvent5 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt5().$action())
                 }
-                NeighborTimerEventSource::TimEvent6(..) => {
+                NeighborTimerEventSource::TimEvent6 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt6().$action())
                 }
-                NeighborTimerEventSource::TimEvent7(..) => {
+                NeighborTimerEventSource::TimEvent7 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt7().$action())
                 }
-                NeighborTimerEventSource::TimEvent8(..) => {
+                NeighborTimerEventSource::TimEvent8 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt8().$action())
                 }
-                NeighborTimerEventSource::TimEvent9(..) => {
+                NeighborTimerEventSource::TimEvent9 { .. } => {
                     tim.$register.modify(|_r, w| w.timevnt9().$action())
                 }
             },
@@ -1980,1208 +2014,3 @@ pub trait HrtimChannel<PSCL> {}
 
 impl<PSCL> HrtimChannel<PSCL> for CH1<PSCL> {}
 impl<PSCL> HrtimChannel<PSCL> for CH2<PSCL> {}
-
-pub enum FaultAction {
-    /// Output never enters fault mode
-    None = 0b00,
-
-    /// Output forced to `active` level on fault
-    ForceActive = 0b01,
-
-    /// Output forced to `inactive` level on fault
-    ForceInactive = 0b10,
-
-    /// The output is floating/tri stated on fault
-    Floating = 0b11,
-}
-
-pub unsafe trait FaultSource: Copy {
-    const ENABLE_BITS: u8;
-}
-
-pub struct SourceBuilder<I> {
-    _input: I,
-    src_bits: u8,
-
-    /// FLTxP
-    is_active_high: bool,
-
-    /// FLTxF[3:0]
-    filter_bits: u8,
-}
-
-impl<I> SourceBuilder<I> {
-    unsafe fn new(input: I, src_bits: u8) -> Self {
-        SourceBuilder {
-            _input: input,
-            src_bits,
-            is_active_high: false,
-            filter_bits: 0b0000,
-        }
-    }
-}
-
-macro_rules! impl_faults {
-    ($(
-        $input:ident => $source:ident:
-            PINS=[($pin:ident, $af:ident), $(($pin_b:ident, $af_b:ident),)*],
-            COMP=$compX:ident, $enable_bits:literal,
-            $fltinrZ:ident, $fltWsrc_0:ident, $fltWsrc_1:ident, $fltWp:ident, $fltWf:ident, $fltWe:ident, $fltWlck:ident,
-    )+) => {$(
-
-        // This should NOT be Copy/Clone
-        pub struct $input {
-            _x: PhantomData<()>
-        }
-
-        #[derive(Copy, Clone)]
-        pub struct $source {
-            _x: PhantomData<()>
-        }
-
-        impl $input {
-            pub fn bind_pin<IM>(self, pin: $pin<gpio::Input<IM>>) -> SourceBuilder<$input> {
-                pin.into_alternate::<$af>();
-                unsafe { SourceBuilder::new(self, 0b00) }
-            }
-
-            $(
-                pub fn bind_pin_b<IM>(self, pin: $pin_b<gpio::Input<IM>>) -> SourceBuilder<$input> {
-                    pin.into_alternate::<$af_b>();
-                    unsafe { SourceBuilder::new(self, 0b00) }
-                }
-            )*
-
-            pub fn bind_comp(self, _comp: &crate::comparator::Comparator<$compX, crate::comparator::Enabled>) -> SourceBuilder<$input> {
-                unsafe { SourceBuilder::new(self, 0b01) }
-            }
-
-            /*pub fn bind_external(?) {
-                SourceBuilder::new(self, 0b10);
-            }*/
-        }
-
-        impl SourceBuilder<$input> {
-            pub fn finalize(self, _control: &mut HrPwmControl) -> $source {
-                let SourceBuilder{ _input, src_bits, is_active_high, filter_bits } = self;
-
-                // Setup fault source
-                unsafe {
-                    let common = &*HRTIM_COMMON::ptr();
-
-                    common.fltinr2.modify(|_r, w| w.$fltWsrc_1().bit(src_bits & 0b10 != 0));
-                    common.$fltinrZ.modify(|_r, w| w
-                        .$fltWsrc_0().bit(src_bits & 0b01 != 0)
-                        .$fltWp().bit(is_active_high)
-                        .$fltWf().bits(filter_bits)
-                        .$fltWe().set_bit() // Enable
-                    );
-
-                    // ... and lock configuration
-                    common.$fltinrZ.modify(|_r, w| w.$fltWlck().set_bit());
-                }
-
-                $source {
-                    _x: PhantomData
-                }
-            }
-
-            pub fn polarity(mut self, polarity: super::Polarity) -> Self {
-                self.is_active_high = polarity == super::Polarity::ActiveHigh;
-                self
-            }
-
-            // TODO: add more settings
-            /* pub fn blanking(?) -> Self */
-
-            pub fn filter(mut self, filter: FaultSamplingFilter) -> Self {
-                self.filter_bits = filter as u8;
-                self
-            }
-        }
-
-        unsafe impl FaultSource for $source {
-            const ENABLE_BITS: u8 = $enable_bits;
-        }
-    )+}
-}
-
-// TODO: Lookup to ensure the alternate function are the same for other devices than stm32g474
-#[cfg(feature = "stm32g474")]
-impl_faults!(
-    FaultInput1 => FaultSource1: PINS=[(PA12, AF13),], COMP=COMP2, 0b000001, fltinr1, flt1src, flt1src_1, flt1p, flt1f, flt1e, flt1lck,
-    FaultInput2 => FaultSource2: PINS=[(PA15, AF13),], COMP=COMP4, 0b000010, fltinr1, flt2src, flt2src_1, flt2p, flt2f, flt2e, flt2lck,
-    FaultInput3 => FaultSource3: PINS=[(PB10, AF13),], COMP=COMP6, 0b000100, fltinr1, flt3src, flt3src_1, flt3p, flt3f, flt3e, flt3lck,
-    FaultInput4 => FaultSource4: PINS=[(PB11, AF13),], COMP=COMP1, 0b001000, fltinr1, flt4src, flt4src_1, flt4p, flt4f, flt4e, flt4lck,
-    FaultInput5 => FaultSource5: PINS=[(PB0, AF13), (PC7, AF3),], COMP=COMP3, 0b010000, fltinr2, flt5src, flt5src_1, flt5p, flt5f, flt5e, flt5lck,
-    FaultInput6 => FaultSource6: PINS=[(PC10, AF13),], COMP=COMP5, 0b100000, fltinr2, flt6src_0, flt6src_1, flt6p, flt6f, flt6e, flt6lck,
-);
-
-pub struct FaultInputs {
-    pub fault_input1: FaultInput1,
-    pub fault_input2: FaultInput2,
-    pub fault_input3: FaultInput3,
-    pub fault_input4: FaultInput4,
-    pub fault_input5: FaultInput5,
-    pub fault_input6: FaultInput6,
-}
-
-pub struct HrPwmControl {
-    _x: PhantomData<()>,
-
-    pub fault_sys: FltMonitorSys,
-    pub fault_1: FltMonitor1,
-    pub fault_2: FltMonitor2,
-    pub fault_3: FltMonitor3,
-    pub fault_4: FltMonitor4,
-    pub fault_5: FltMonitor5,
-    pub fault_6: FltMonitor6,
-}
-
-/// The divsion ratio between f_hrtim and the fault signal sampling clock for digital filters
-pub enum FaultSamplingClkDiv {
-    /// No division
-    ///
-    /// fault signal sampling clock f_flts = f_hrtim
-    None = 0b00,
-
-    /// 1/2
-    ///
-    /// fault signal sampling clock f_flts = f_hrtim / 2
-    Two = 0b01,
-
-    /// 1/4
-    ///
-    /// fault signal sampling clock f_flts = f_hrtim / 4
-    Four = 0b10,
-
-    /// 1/8
-    ///
-    /// fault signal sampling clock f_flts = f_hrtim / 8
-    Eight = 0b11,
-}
-
-pub enum FaultSamplingFilter {
-    /// No filtering, fault acts asynchronously
-    ///
-    /// Note that this bypasses any f_flts (FaultSamplingClkDiv)
-    None = 0b0000,
-
-    /// Sample directly at rate f_hrtim, with a count of 2
-    ///
-    /// Note that this bypasses: any f_flts (FaultSamplingClkDiv)
-    HrtimN2 = 0b0001,
-
-    /// Sample directly at rate f_hrtim, with a count of 4
-    ///
-    /// Note that this bypasses any f_flts (FaultSamplingClkDiv)
-    HrtimN4 = 0b0010,
-
-    /// Sample directly at rate f_hrtim, with a count of 8
-    ///
-    /// Note that this bypasses any f_flts (FaultSamplingClkDiv)
-    HrtimN8 = 0b0011,
-
-    /// Sample at rate f_flts / 2, with a count of 6
-    FltsDiv2N6 = 0b0100,
-
-    /// Sample at rate f_flts / 2, with a count of 8
-    FltsDiv2N8 = 0b0101,
-
-    /// Sample at rate f_flts / 4, with a count of 6
-    FltsDiv4N6 = 0b0110,
-
-    /// Sample at rate f_flts / 4, with a count of 8
-    FltsDiv4N8 = 0b0111,
-
-    /// Sample at rate f_flts / 8, with a count of 6
-    FltsDiv8N6 = 0b1000,
-
-    /// Sample at rate f_flts / 8, with a count of 8
-    FltsDiv8N8 = 0b1001,
-
-    /// Sample at rate f_flts / 16, with a count of 5
-    FltsDiv16N5 = 0b1010,
-
-    /// Sample at rate f_flts / 16, with a count of 6
-    FltsDiv16N6 = 0b1011,
-
-    /// Sample at rate f_flts / 16, with a count of 8
-    FltsDiv16N8 = 0b1100,
-
-    /// Sample at rate f_flts / 32, with a count of 5
-    FltsDiv32N5 = 0b1101,
-
-    /// Sample at rate f_flts / 32, with a count of 6
-    FltsDiv32N6 = 0b1110,
-
-    /// Sample at rate f_flts / 32, with a count of 8
-    FltsDiv32N8 = 0b1111,
-}
-
-pub trait HrControltExt {
-    fn hr_control(self, _rcc: &mut Rcc) -> HrTimOngoingCalibration;
-}
-
-impl HrControltExt for HRTIM_COMMON {
-    fn hr_control(self, _rcc: &mut Rcc) -> HrTimOngoingCalibration {
-        let common = unsafe { &*HRTIM_COMMON::ptr() };
-
-        unsafe {
-            let rcc_ptr = &*RCC::ptr();
-
-            HRTIM_COMMON::enable(rcc_ptr);
-            HRTIM_COMMON::reset(rcc_ptr);
-        }
-
-        // Start calibration procedure
-        common
-            .dllcr
-            .write(|w| w.cal().set_bit().calen().clear_bit());
-
-        HrTimOngoingCalibration {
-            adc_trigger1_bits: 0,
-            adc_trigger2_bits: 0,
-            adc_trigger3_bits: 0,
-            adc_trigger4_bits: 0,
-
-            adc_trigger5_bits: 0,
-            adc_trigger6_bits: 0,
-            adc_trigger7_bits: 0,
-            adc_trigger8_bits: 0,
-            adc_trigger9_bits: 0,
-            adc_trigger10_bits: 0,
-
-            adc_trigger1_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger2_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger3_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger4_postscaler: AdcTriggerPostscaler::None,
-
-            adc_trigger5_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger6_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger7_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger8_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger9_postscaler: AdcTriggerPostscaler::None,
-            adc_trigger10_postscaler: AdcTriggerPostscaler::None,
-
-            divider: FaultSamplingClkDiv::None,
-        }
-    }
-}
-
-pub struct HrTimOngoingCalibration {
-    adc_trigger1_bits: u32,
-    adc_trigger2_bits: u32,
-    adc_trigger3_bits: u32,
-    adc_trigger4_bits: u32,
-
-    adc_trigger5_bits: u8,
-    adc_trigger6_bits: u8,
-    adc_trigger7_bits: u8,
-    adc_trigger8_bits: u8,
-    adc_trigger9_bits: u8,
-    adc_trigger10_bits: u8,
-
-    adc_trigger1_postscaler: AdcTriggerPostscaler,
-    adc_trigger2_postscaler: AdcTriggerPostscaler,
-    adc_trigger3_postscaler: AdcTriggerPostscaler,
-    adc_trigger4_postscaler: AdcTriggerPostscaler,
-
-    adc_trigger5_postscaler: AdcTriggerPostscaler,
-    adc_trigger6_postscaler: AdcTriggerPostscaler,
-    adc_trigger7_postscaler: AdcTriggerPostscaler,
-    adc_trigger8_postscaler: AdcTriggerPostscaler,
-    adc_trigger9_postscaler: AdcTriggerPostscaler,
-    adc_trigger10_postscaler: AdcTriggerPostscaler,
-
-    divider: FaultSamplingClkDiv,
-}
-
-impl HrTimOngoingCalibration {
-    /// SAFETY: Calibration needs to be done before calling this
-    unsafe fn init(self) -> (HrPwmControl, FaultInputs) {
-        use Adc13Trigger as Ad13T;
-        use Adc24Trigger as Ad24T;
-
-        let common = unsafe { &*HRTIM_COMMON::ptr() };
-
-        let Self {
-            adc_trigger1_bits: ad1_bits,
-            adc_trigger2_bits: ad2_bits,
-            adc_trigger3_bits: ad3_bits,
-            adc_trigger4_bits: ad4_bits,
-
-            adc_trigger5_bits: ad5_bits,
-            adc_trigger6_bits: ad6_bits,
-            adc_trigger7_bits: ad7_bits,
-            adc_trigger8_bits: ad8_bits,
-            adc_trigger9_bits: ad9_bits,
-            adc_trigger10_bits: ad10_bits,
-
-            adc_trigger1_postscaler,
-            adc_trigger2_postscaler,
-            adc_trigger3_postscaler,
-            adc_trigger4_postscaler,
-
-            adc_trigger5_postscaler,
-            adc_trigger6_postscaler,
-            adc_trigger7_postscaler,
-            adc_trigger8_postscaler,
-            adc_trigger9_postscaler,
-            adc_trigger10_postscaler,
-
-            divider,
-        } = self;
-
-        unsafe {
-            // Enable periodic calibration
-            // with f_hrtim at 170MHz, these settings leads to
-            // a period of about 6.2ms
-            common
-                .dllcr
-                .modify(|_r, w| w.calrte().bits(0b00).cal().set_bit().calen().clear_bit());
-            common.fltinr2.write(|w| w.fltsd().bits(divider as u8));
-
-            common.adc1r.write(|w| {
-                w.eper()
-                    .bit(ad1_bits | Ad13T::TimEPeriod as u32 != 0)
-                    .ec4()
-                    .bit(ad1_bits | Ad13T::TimECmp4 as u32 != 0)
-                    .ec3()
-                    .bit(ad1_bits | Ad13T::TimECmp3 as u32 != 0)
-                    //.frst()
-                    .dper()
-                    .bit(ad1_bits | Ad13T::TimDPeriod as u32 != 0)
-                    .dc4()
-                    .bit(ad1_bits | Ad13T::TimDCmp4 as u32 != 0)
-                    .dc3()
-                    .bit(ad1_bits | Ad13T::TimDCmp3 as u32 != 0)
-                    .fper()
-                    .bit(ad1_bits | Ad13T::TimFPeriod as u32 != 0)
-                    .cper()
-                    .bit(ad1_bits | Ad13T::TimCPeriod as u32 != 0)
-                    .cc4()
-                    .bit(ad1_bits | Ad13T::TimCCmp4 as u32 != 0)
-                    .cc3()
-                    .bit(ad1_bits | Ad13T::TimCCmp3 as u32 != 0)
-                    .fc4()
-                    .bit(ad1_bits | Ad13T::TimFCmp4 as u32 != 0)
-                    //.brst()
-                    .bper()
-                    .bit(ad1_bits | Ad13T::TimBPeriod as u32 != 0)
-                    .bc4()
-                    .bit(ad1_bits | Ad13T::TimBCmp4 as u32 != 0)
-                    .bc3()
-                    .bit(ad1_bits | Ad13T::TimBCmp3 as u32 != 0)
-                    .fc3()
-                    .bit(ad1_bits | Ad13T::TimFCmp3 as u32 != 0)
-                    //.arst()
-                    .aper()
-                    .bit(ad1_bits | Ad13T::TimAPeriod as u32 != 0)
-                    .ac4()
-                    .bit(ad1_bits | Ad13T::TimACmp4 as u32 != 0)
-                    .ac3()
-                    .bit(ad1_bits | Ad13T::TimACmp3 as u32 != 0)
-                    .fc2()
-                    .bit(ad1_bits | Ad13T::TimFCmp2 as u32 != 0)
-                    //.eev5().bit(ad1_bits | Ad13T::_ as u32)
-                    //.eev4().bit(ad1_bits | Ad13T::_ as u32)
-                    //.eev3().bit(ad1_bits | Ad13T::_ as u32)
-                    //.eev2().bit(ad1_bits | Ad13T::_ as u32)
-                    //.eev1().bit(ad1_bits | Ad13T::_ as u32)
-                    .mper()
-                    .bit(ad1_bits | Ad13T::MasterPeriod as u32 != 0)
-                    .mc4()
-                    .bit(ad1_bits | Ad13T::MasterCmp4 as u32 != 0)
-                    .mc3()
-                    .bit(ad1_bits | Ad13T::MasterCmp3 as u32 != 0)
-                    .mc2()
-                    .bit(ad1_bits | Ad13T::MasterCmp2 as u32 != 0)
-                    .mc1()
-                    .bit(ad1_bits | Ad13T::MasterCmp1 as u32 != 0)
-            });
-
-            common.adc3r.write(|w| {
-                w.eper()
-                    .bit(ad3_bits | Ad13T::TimEPeriod as u32 != 0)
-                    .ec4()
-                    .bit(ad3_bits | Ad13T::TimECmp4 as u32 != 0)
-                    .ec3()
-                    .bit(ad3_bits | Ad13T::TimECmp3 as u32 != 0)
-                    //.frst()
-                    .dper()
-                    .bit(ad3_bits | Ad13T::TimDPeriod as u32 != 0)
-                    .dc4()
-                    .bit(ad3_bits | Ad13T::TimDCmp4 as u32 != 0)
-                    .dc3()
-                    .bit(ad3_bits | Ad13T::TimDCmp3 as u32 != 0)
-                    .fper()
-                    .bit(ad3_bits | Ad13T::TimFPeriod as u32 != 0)
-                    .cper()
-                    .bit(ad3_bits | Ad13T::TimCPeriod as u32 != 0)
-                    .cc4()
-                    .bit(ad3_bits | Ad13T::TimCCmp4 as u32 != 0)
-                    .cc3()
-                    .bit(ad3_bits | Ad13T::TimCCmp3 as u32 != 0)
-                    .fc4()
-                    .bit(ad3_bits | Ad13T::TimFCmp4 as u32 != 0)
-                    //.brst()
-                    .bper()
-                    .bit(ad3_bits | Ad13T::TimBPeriod as u32 != 0)
-                    .bc4()
-                    .bit(ad3_bits | Ad13T::TimBCmp4 as u32 != 0)
-                    .bc3()
-                    .bit(ad3_bits | Ad13T::TimBCmp3 as u32 != 0)
-                    .fc3()
-                    .bit(ad3_bits | Ad13T::TimFCmp3 as u32 != 0)
-                    //.arst()
-                    .aper()
-                    .bit(ad3_bits | Ad13T::TimAPeriod as u32 != 0)
-                    .ac4()
-                    .bit(ad3_bits | Ad13T::TimACmp4 as u32 != 0)
-                    .ac3()
-                    .bit(ad3_bits | Ad13T::TimACmp3 as u32 != 0)
-                    .fc2()
-                    .bit(ad3_bits | Ad13T::TimFCmp2 as u32 != 0)
-                    //.eev5().bit(ad3_bits | Ad13T::_ as u32)
-                    //.eev4().bit(ad3_bits | Ad13T::_ as u32)
-                    //.eev3().bit(ad3_bits | Ad13T::_ as u32)
-                    //.eev2().bit(ad3_bits | Ad13T::_ as u32)
-                    //.eev1().bit(ad3_bits | Ad13T::_ as u32)
-                    .mper()
-                    .bit(ad3_bits | Ad13T::MasterPeriod as u32 != 0)
-                    .mc4()
-                    .bit(ad3_bits | Ad13T::MasterCmp4 as u32 != 0)
-                    .mc3()
-                    .bit(ad3_bits | Ad13T::MasterCmp3 as u32 != 0)
-                    .mc2()
-                    .bit(ad3_bits | Ad13T::MasterCmp2 as u32 != 0)
-                    .mc1()
-                    .bit(ad3_bits | Ad13T::MasterCmp1 as u32 != 0)
-            });
-
-            common.adc2r.write(|w| {
-                w
-                    //.erst()
-                    .ec4()
-                    .bit(ad2_bits | Ad24T::TimECmp4 as u32 != 0)
-                    .ec3()
-                    .bit(ad2_bits | Ad24T::TimECmp3 as u32 != 0)
-                    .ec2()
-                    .bit(ad2_bits | Ad24T::TimECmp2 as u32 != 0)
-                    //.drst().bit(ad2_bits | Ad24T::_ as u32 != 0)
-                    .dper()
-                    .bit(ad2_bits | Ad24T::TimDPeriod as u32 != 0)
-                    .dc4()
-                    .bit(ad2_bits | Ad24T::TimDCmp4 as u32 != 0)
-                    .fper()
-                    .bit(ad2_bits | Ad24T::TimFPeriod as u32 != 0)
-                    .dc2()
-                    .bit(ad2_bits | Ad24T::TimDCmp2 as u32 != 0)
-                    //.crst().bit(ad2_bits | Ad24T::_ as u32 != 0)
-                    .cper()
-                    .bit(ad2_bits | Ad24T::TimCPeriod as u32 != 0)
-                    .cc4()
-                    .bit(ad2_bits | Ad24T::TimCCmp4 as u32 != 0)
-                    .fc4()
-                    .bit(ad2_bits | Ad24T::TimFCmp4 as u32 != 0)
-                    .cc2()
-                    .bit(ad2_bits | Ad24T::TimCCmp2 as u32 != 0)
-                    .bper()
-                    .bit(ad2_bits | Ad24T::TimBPeriod as u32 != 0)
-                    .bc4()
-                    .bit(ad2_bits | Ad24T::TimBCmp4 as u32 != 0)
-                    .fc3()
-                    .bit(ad2_bits | Ad24T::TimFCmp3 as u32 != 0)
-                    .bc2()
-                    .bit(ad2_bits | Ad24T::TimBCmp2 as u32 != 0)
-                    .aper()
-                    .bit(ad2_bits | Ad24T::TimAPeriod as u32 != 0)
-                    .ac4()
-                    .bit(ad2_bits | Ad24T::TimACmp4 as u32 != 0)
-                    .fc2()
-                    .bit(ad2_bits | Ad24T::TimFCmp2 as u32 != 0)
-                    .ac2()
-                    .bit(ad2_bits | Ad24T::TimACmp2 as u32 != 0)
-                    //.eev10()
-                    //.eev9()
-                    //.eev8()
-                    //.eev7()
-                    //.eev6()
-                    .mper()
-                    .bit(ad2_bits | Ad24T::MasterPeriod as u32 != 0)
-                    .mc4()
-                    .bit(ad2_bits | Ad24T::MasterCmp4 as u32 != 0)
-                    .mc3()
-                    .bit(ad2_bits | Ad24T::MasterCmp3 as u32 != 0)
-                    .mc2()
-                    .bit(ad2_bits | Ad24T::MasterCmp2 as u32 != 0)
-                    .mc1()
-                    .bit(ad2_bits | Ad24T::MasterCmp1 as u32 != 0)
-            });
-
-            common.adc4r.write(|w| {
-                w
-                    //.erst()
-                    .ec4()
-                    .bit(ad4_bits | Ad24T::TimECmp4 as u32 != 0)
-                    .ec3()
-                    .bit(ad4_bits | Ad24T::TimECmp3 as u32 != 0)
-                    .ec2()
-                    .bit(ad4_bits | Ad24T::TimECmp2 as u32 != 0)
-                    //.drst().bit(ad4_bits | Ad24T::_ as u32 != 0)
-                    .dper()
-                    .bit(ad4_bits | Ad24T::TimDPeriod as u32 != 0)
-                    .dc4()
-                    .bit(ad4_bits | Ad24T::TimDCmp4 as u32 != 0)
-                    .fper()
-                    .bit(ad4_bits | Ad24T::TimFPeriod as u32 != 0)
-                    .dc2()
-                    .bit(ad4_bits | Ad24T::TimDCmp2 as u32 != 0)
-                    //.crst().bit(ad4_bits | Ad24T::_ as u32 != 0)
-                    .cper()
-                    .bit(ad4_bits | Ad24T::TimCPeriod as u32 != 0)
-                    .cc4()
-                    .bit(ad4_bits | Ad24T::TimCCmp4 as u32 != 0)
-                    .fc4()
-                    .bit(ad4_bits | Ad24T::TimFCmp4 as u32 != 0)
-                    .cc2()
-                    .bit(ad4_bits | Ad24T::TimCCmp2 as u32 != 0)
-                    .bper()
-                    .bit(ad4_bits | Ad24T::TimBPeriod as u32 != 0)
-                    .bc4()
-                    .bit(ad4_bits | Ad24T::TimBCmp4 as u32 != 0)
-                    .fc3()
-                    .bit(ad4_bits | Ad24T::TimFCmp3 as u32 != 0)
-                    .bc2()
-                    .bit(ad4_bits | Ad24T::TimBCmp2 as u32 != 0)
-                    .aper()
-                    .bit(ad4_bits | Ad24T::TimAPeriod as u32 != 0)
-                    .ac4()
-                    .bit(ad4_bits | Ad24T::TimACmp4 as u32 != 0)
-                    .fc2()
-                    .bit(ad4_bits | Ad24T::TimFCmp2 as u32 != 0)
-                    .ac2()
-                    .bit(ad4_bits | Ad24T::TimACmp2 as u32 != 0)
-                    //.eev10()
-                    //.eev9()
-                    //.eev8()
-                    //.eev7()
-                    //.eev6()
-                    .mper()
-                    .bit(ad4_bits | Ad24T::MasterPeriod as u32 != 0)
-                    .mc4()
-                    .bit(ad4_bits | Ad24T::MasterCmp4 as u32 != 0)
-                    .mc3()
-                    .bit(ad4_bits | Ad24T::MasterCmp3 as u32 != 0)
-                    .mc2()
-                    .bit(ad4_bits | Ad24T::MasterCmp2 as u32 != 0)
-                    .mc1()
-                    .bit(ad4_bits | Ad24T::MasterCmp1 as u32 != 0)
-            });
-
-            common.adcer.write(|w| {
-                w.adc5trg()
-                    .variant(ad5_bits)
-                    .adc6trg()
-                    .variant(ad6_bits)
-                    .adc7trg()
-                    .variant(ad7_bits)
-                    .adc8trg()
-                    .variant(ad8_bits)
-                    .adc9trg()
-                    .variant(ad9_bits)
-                    .adc10trg()
-                    .variant(ad10_bits)
-            });
-
-            common.adcps1.write(|w| {
-                w.adc1psc()
-                    .bits(adc_trigger1_postscaler as u8)
-                    .adc2psc()
-                    .bits(adc_trigger2_postscaler as u8)
-                    .adc3psc()
-                    .bits(adc_trigger3_postscaler as u8)
-                    .adc4psc()
-                    .bits(adc_trigger4_postscaler as u8)
-                    .adc5psc()
-                    .bits(adc_trigger5_postscaler as u8)
-            });
-
-            common.adcps2.write(|w| {
-                w.adc6psc()
-                    .bits(adc_trigger6_postscaler as u8)
-                    .adc7psc()
-                    .bits(adc_trigger7_postscaler as u8)
-                    .adc8psc()
-                    .bits(adc_trigger8_postscaler as u8)
-                    .adc9psc()
-                    .bits(adc_trigger9_postscaler as u8)
-                    .adc10psc()
-                    .bits(adc_trigger10_postscaler as u8)
-            });
-
-            // TODO: Adc trigger 5-10
-        }
-
-        (
-            HrPwmControl {
-                _x: PhantomData,
-                fault_sys: FltMonitorSys { _x: PhantomData },
-                fault_1: FltMonitor1 { _x: PhantomData },
-                fault_2: FltMonitor2 { _x: PhantomData },
-                fault_3: FltMonitor3 { _x: PhantomData },
-                fault_4: FltMonitor4 { _x: PhantomData },
-                fault_5: FltMonitor5 { _x: PhantomData },
-                fault_6: FltMonitor6 { _x: PhantomData },
-            },
-            FaultInputs {
-                fault_input1: FaultInput1 { _x: PhantomData },
-                fault_input2: FaultInput2 { _x: PhantomData },
-                fault_input3: FaultInput3 { _x: PhantomData },
-                fault_input4: FaultInput4 { _x: PhantomData },
-                fault_input5: FaultInput5 { _x: PhantomData },
-                fault_input6: FaultInput6 { _x: PhantomData },
-            },
-        )
-    }
-
-    pub fn wait_for_calibration(self) -> (HrPwmControl, FaultInputs) {
-        let common = unsafe { &*HRTIM_COMMON::ptr() };
-        while common.isr.read().dllrdy().bit_is_clear() {
-            // Wait until ready
-        }
-
-        // Calibration is now done, it is safe to continue
-        unsafe { self.init() }
-    }
-
-    pub fn enable_adc_trigger1_source(mut self, trigger: Adc13Trigger) -> Self {
-        self.adc_trigger1_bits |= trigger as u32;
-        self
-    }
-
-    pub fn enable_adc_trigger2_source(mut self, trigger: Adc24Trigger) -> Self {
-        self.adc_trigger2_bits |= trigger as u32;
-        self
-    }
-
-    pub fn enable_adc_trigger3_source(mut self, trigger: Adc13Trigger) -> Self {
-        self.adc_trigger3_bits |= trigger as u32;
-        self
-    }
-
-    pub fn enable_adc_trigger4_source(mut self, trigger: Adc24Trigger) -> Self {
-        self.adc_trigger4_bits |= trigger as u32;
-        self
-    }
-
-    pub fn enable_adc_trigger5_source(mut self, trigger: Adc579Trigger) -> Self {
-        self.adc_trigger5_bits = trigger as u8;
-        self
-    }
-
-    pub fn enable_adc_trigger6_source(mut self, trigger: Adc6810Trigger) -> Self {
-        self.adc_trigger6_bits = trigger as u8;
-        self
-    }
-
-    pub fn enable_adc_trigger7_source(mut self, trigger: Adc579Trigger) -> Self {
-        self.adc_trigger7_bits = trigger as u8;
-        self
-    }
-
-    pub fn enable_adc_trigger8_source(mut self, trigger: Adc6810Trigger) -> Self {
-        self.adc_trigger8_bits = trigger as u8;
-        self
-    }
-
-    pub fn enable_adc_trigger9_source(mut self, trigger: Adc579Trigger) -> Self {
-        self.adc_trigger9_bits = trigger as u8;
-        self
-    }
-
-    pub fn enable_adc_trigger10_source(mut self, trigger: Adc6810Trigger) -> Self {
-        self.adc_trigger10_bits = trigger as u8;
-        self
-    }
-
-    pub fn set_adc1_trigger_psc(mut self, post_scaler: AdcTriggerPostscaler) -> Self {
-        self.adc_trigger1_postscaler = post_scaler;
-        self
-    }
-
-    pub fn set_adc2_trigger_psc(mut self, post_scaler: AdcTriggerPostscaler) -> Self {
-        self.adc_trigger2_postscaler = post_scaler;
-        self
-    }
-
-    pub fn set_adc3_trigger_psc(mut self, post_scaler: AdcTriggerPostscaler) -> Self {
-        self.adc_trigger3_postscaler = post_scaler;
-        self
-    }
-
-    pub fn set_adc4_trigger_psc(mut self, post_scaler: AdcTriggerPostscaler) -> Self {
-        self.adc_trigger4_postscaler = post_scaler;
-        self
-    }
-
-    pub fn set_fault_sampling_division(mut self, divider: FaultSamplingClkDiv) -> Self {
-        self.divider = divider;
-        self
-    }
-
-    // TODO: Adc trigger 5-10
-}
-
-pub enum Adc13Trigger {
-    /// bit 31 ADCxTEPER - Trigger on HRTIM_TIME period
-    TimEPeriod = 1 << 31,
-
-    /// bit 30 ADCxTEC4 - Trigger on HRTIM_TIME compare match for compare register 4
-    TimECmp4 = 1 << 30,
-
-    /// bit 29 ADCxTEC3 - Trigger on HRTIM_TIME compare match for compare register 3
-    TimECmp3 = 1 << 29,
-
-    /// bit 28 ADCxTFRST - Trigger on HRTIM_TIMF reset or counter roll-over
-    TimFRst = 1 << 28,
-
-    /// bit 27 ADCxTDPER - Trigger on HRTIM_TIMD period
-    TimDPeriod = 1 << 27,
-
-    /// bit 26 ADCxTDC4 - Trigger on HRTIM_TIMD compare match for compare register 4
-    TimDCmp4 = 1 << 26,
-
-    /// bit 25 ADCxTDC3 - Trigger on HRTIM_TIMD compare match for compare register 3
-    TimDCmp3 = 1 << 25,
-
-    /// bit 24 ADCxTFPER - Trigger on HRTIM_TIMF period
-    TimFPeriod = 1 << 24,
-
-    /// bit 23 ADCxTCPER - Trigger on HRTIM_TIMC period
-    TimCPeriod = 1 << 23,
-
-    /// bit 22 ADCxTCC4 - Trigger on HRTIM_TIMC compare match for compare register 4
-    TimCCmp4 = 1 << 22,
-
-    /// bit 21 ADCxTCC3 - Trigger on HRTIM_TIMC compare match for compare register 3
-    TimCCmp3 = 1 << 21,
-
-    /// bit 20 ADCxTFC4 - Trigger on HRTIM_TIMF compare match for compare register 4
-    TimFCmp4 = 1 << 20,
-
-    /// bit 19 ADCxTBRST - Trigger on HRTIM_TIMB reset or counter roll-over
-    TimBRst = 1 << 19,
-
-    /// bit 18 ADCxTBPER - Trigger on HRTIM_TIMB period
-    TimBPeriod = 1 << 18,
-
-    /// bit 17 ADCxTBC4 - Trigger on HRTIM_TIMB compare match for compare register 4
-    TimBCmp4 = 1 << 17,
-
-    /// bit 16 ADCxTBC3 - Trigger on HRTIM_TIMB compare match for compare register 3
-    TimBCmp3 = 1 << 16,
-
-    /// bit 15 ADCxTFC3 - Trigger on HRTIM_TIMF compare match for compare register 3
-    TimFCmp3 = 1 << 15,
-
-    /// bit 14 ADCxTARST - Trigger on HRTIM_TIMA reset or counter roll-over
-    TimARst = 1 << 14,
-
-    /// bit 13 ADCxTAPER - Trigger on HRTIM_TIMA period
-    TimAPeriod = 1 << 13,
-
-    /// bit 12 ADCxTAC4 - Trigger on HRTIM_TIMA compare match for compare register 4
-    TimACmp4 = 1 << 12,
-
-    /// bit 11 ADCxTAC3 - Trigger on HRTIM_TIMA compare match for compare register 3
-    TimACmp3 = 1 << 11,
-
-    /// bit 10 ADCxTFC2 - Trigger on HRTIM_TIMF compare match for compare register 2
-    TimFCmp2 = 1 << 10,
-
-    // /// bit 9 ADCxEEV5
-    // _ = 1 << 9,
-
-    // /// bit 8 ADCxEEV4
-    // _ = 1 << 8,
-
-    // /// bit 7 ADCxEEV3
-    // _ = 1 << 7,
-
-    // /// bit 6 ADCxEEV2
-    // _ = 1 << 6,
-    /// bit 5 ADCxEEV1
-    // _ = 1 << 5,
-
-    /// bit 4 ADCxMPER - Trigger on HRTIM_MASTER period
-    MasterPeriod = 1 << 4,
-
-    /// bit 3 ADCxMC4 - Trigger on HRTIM_MASTER compare match for compare register 4
-    MasterCmp4 = 1 << 3,
-
-    /// bit 2 ADCxMC3 - Trigger on HRTIM_MASTER compare match for compare register 3
-    MasterCmp3 = 1 << 2,
-
-    /// bit 1 ADCxMC2 - Trigger on HRTIM_MASTER compare match for compare register 2
-    MasterCmp2 = 1 << 1,
-
-    /// bit 0 ADCxMC1 - Trigger on HRTIM_MASTER compare match for compare register 1
-    MasterCmp1 = 1 << 0,
-}
-
-pub enum AdcTriggerPostscaler {
-    None = 0,
-    Div2 = 1,
-    Div3 = 2,
-    Div4 = 3,
-    Div5 = 4,
-    Div6 = 5,
-    Div7 = 6,
-    Div8 = 7,
-    Div9 = 8,
-    Div10 = 9,
-    Div11 = 10,
-    Div12 = 11,
-    Div13 = 12,
-    Div14 = 13,
-    Div15 = 14,
-    Div16 = 15,
-    Div17 = 16,
-    Div18 = 17,
-    Div19 = 18,
-    Div20 = 19,
-    Div21 = 20,
-    Div22 = 21,
-    Div23 = 22,
-    Div24 = 23,
-    Div25 = 24,
-    Div26 = 25,
-    Div27 = 26,
-    Div28 = 27,
-    Div29 = 28,
-    Div30 = 29,
-    Div31 = 30,
-    Div32 = 31,
-}
-
-pub enum Adc24Trigger {
-    /// bit 31 ADCxTERST - Trigger on HRTIM_TIME reset or counter roll-over
-    TimERst = 1 << 31,
-
-    /// bit 30 ADCxTEC4 - Trigger on HRTIM_TIME compare match for compare register 4
-    TimECmp4 = 1 << 30,
-
-    /// bit 29 ADCxTEC3 - Trigger on HRTIM_TIME compare match for compare register 3
-    TimECmp3 = 1 << 29,
-
-    /// bit 28 ADCxTEC2 - Trigger on HRTIM_TIME compare match for compare register 2
-    TimECmp2 = 1 << 28,
-
-    /// bit 27 ADCxTDRST - Trigger on HRTIM_TIMD reset or counter roll-over
-    TimDRst = 1 << 27,
-
-    /// bit 26 ADCxTDPER - Trigger on HRTIM_TIMD period
-    TimDPeriod = 1 << 26,
-
-    /// bit 25 ADCxTDC4 - Trigger on HRTIM_TIMD compare match for compare register 4
-    TimDCmp4 = 1 << 25,
-
-    /// bit 24 ADCxTFPER - Trigger on HRTIM_TIMF period
-    TimFPeriod = 1 << 24,
-
-    /// bit 23 ADCxTDC2 - Trigger on HRTIM_TIMD compare match for compare register 2
-    TimDCmp2 = 1 << 23,
-
-    /// bit 22 ADCxTCRST - Trigger on HRTIM_TIMC reset or counter roll-over
-    TimCRst = 1 << 22,
-
-    /// bit 21 ADCxTCPER - Trigger on HRTIM_TIMC period
-    TimCPeriod = 1 << 21,
-
-    /// bit 20 ADCxTCC4 - Trigger on HRTIM_TIMC compare match for compare register 4
-    TimCCmp4 = 1 << 20,
-
-    /// bit 19 ADCxTFC4 - Trigger on HRTIM_TIMF compare match for compare register 4
-    TimFCmp4 = 1 << 19,
-
-    /// bit 18 ADCxTCC2 - Trigger on HRTIM_TIMC compare match for compare register 2
-    TimCCmp2 = 1 << 18,
-
-    /// bit 17 ADCxTBPER - Trigger on HRTIM_TIMB period
-    TimBPeriod = 1 << 17,
-
-    /// bit 16 ADCxTBC4  - Trigger on HRTIM_TIMB compare match for compare register 4
-    TimBCmp4 = 1 << 16,
-
-    /// bit 15 ADCxTFC3 - Trigger on HRTIM_TIMF compare match for compare register 3
-    TimFCmp3 = 1 << 15,
-
-    /// bit 14 ADCxTBC2 - Trigger on HRTIM_TIMB compare match for compare register 2
-    TimBCmp2 = 1 << 14,
-
-    /// bit 13 ADCxTAPER - Trigger on HRTIM_TIMA period
-    TimAPeriod = 1 << 13,
-
-    /// bit 12 ADCxTAC4 - Trigger on HRTIM_TIMA compare match for compare register 4
-    TimACmp4 = 1 << 12,
-
-    /// bit 11 ADCxTFC2 - Trigger on HRTIM_TIMF compare match for compare register 2
-    TimFCmp2 = 1 << 11,
-
-    /// bit 10 ADCxTAC2 - Trigger on HRTIM_TIMA compare match for compare register 2
-    TimACmp2 = 1 << 10,
-
-    // /// bit 9 ADCxEEV10
-    // _ = 1 << 9,
-
-    // /// bit 8 ADCxEEV9
-    // _ = 1 << 8,
-
-    // /// bit 7 ADCxEEV8
-    // _ = 1 << 7,
-
-    // /// bit 6 ADCxEEV7
-    // _ = 1 << 6,
-
-    // /// bit 5 ADCxEEV6
-    // _ = 1 << 5,
-    /// bit 4 ADCxMPER - Trigger on HRTIM_MASTER period
-    MasterPeriod = 1 << 4,
-
-    /// bit 3 ADCxMC4 - Trigger on HRTIM_MASTER compare match for compare register 4
-    MasterCmp4 = 1 << 3,
-
-    /// bit 2 ADCxMC3 - Trigger on HRTIM_MASTER compare match for compare register 3
-    MasterCmp3 = 1 << 2,
-
-    /// bit 1 ADCxMC2 - Trigger on HRTIM_MASTER compare match for compare register 2
-    MasterCmp2 = 1 << 1,
-
-    /// bit 0 ADCxMC1 - Trigger on HRTIM_MASTER compare match for compare register 1
-    MasterCmp1 = 1 << 0,
-}
-
-pub enum Adc579Trigger {
-    // /// Trigger on F reset and counter roll-over
-    // _ = 31
-    /// Trigger on HRTIM_TIMF period
-    TimFPeriod = 30,
-
-    /// Trigger on HRTIM_TIMF compare match for compare register 4
-    TimFCmp4 = 29,
-
-    /// Trigger on HRTIM_TIMF compare match for compare register 3
-    TimFCmp3 = 28,
-
-    /// Trigger on HRTIM_TIMF compare match for compare register 2
-    TimFCmp2 = 27,
-
-    /// Trigger on HRTIM_TIME period
-    TimEPeriod = 26,
-
-    /// Trigger on HRTIM_TIME compare match for compare register 4
-    TimECmp4 = 25,
-
-    /// Trigger on HRTIM_TIME compare match for compare register 3
-    TimECmp3 = 24,
-
-    /// Trigger on HRTIM_TIMD period
-    TimDPeriod = 23,
-
-    /// Trigger on HRTIM_TIMD compare match for compare register 4
-    TimDCmp4 = 22,
-
-    /// Trigger on HRTIM_TIMD compare match for compare register 3
-    TimDCmp3 = 21,
-
-    /// Trigger on HRTIM_TIMC period
-    TimCPeriod = 20,
-
-    /// Trigger on HRTIM_TIMC compare match for compare register 4
-    TimCCmp4 = 19,
-
-    /// Trigger on HRTIM_TIMC compare match for compare register 3
-    TimCCmp3 = 18,
-
-    // /// Trigger on B reset and counter roll-over
-    // _ = 17
-    /// Trigger on HRTIM_TIMB period
-    TimBPeriod = 16,
-
-    /// Trigger on HRTIM_TIMB compare match for compare register 4
-    TimBCmp4 = 15,
-
-    /// Trigger on HRTIM_TIMB compare match for compare register 3
-    TimBCmp3 = 14,
-
-    // /// Trigger on A reset and counter roll-over
-    // _ = 13
-    /// Trigger on HRTIM_TIMA period
-    TimAPeriod = 12,
-
-    /// Trigger on HRTIM_TIMA compare match for compare register 4
-    TimACmp4 = 11,
-
-    /// Trigger on HRTIM_TIMA compare match for compare register 3
-    TimACmp3 = 10,
-
-    // /// Trigger on EEV5
-    // _ = 9,
-
-    // ///  Trigger on EEV4
-    // _ = 8,
-
-    // ///  Trigger on EEV3
-    // _ = 7,
-
-    // ///  Trigger on EEV2
-    // _ = 6,
-
-    // ///  Trigger on EEV1
-    // _ = 5,
-    /// Trigger on HRTIM_MASTER period
-    MasterPeriod = 4,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 4
-    MasterCmp4 = 3,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 3
-    MasterCmp3 = 2,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 2
-    MasterCmp2 = 1,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 1
-    MasterCmp1 = 0,
-}
-
-pub enum Adc6810Trigger {
-    /// Trigger on HRTIM_TIMF period
-    TimFPeriod = 31,
-
-    /// Trigger on HRTIM_TIMF compare match for compare register 4
-    TimFCmp4 = 30,
-
-    /// Trigger on HRTIM_TIMF compare match for compare register 3
-    TimFCmp3 = 29,
-
-    /// Trigger on HRTIM_TIMF compare match for compare register 2
-    TimFCmp2 = 28,
-
-    // /// Trigger on E reset and counter roll-over
-    // _ = 27
-    /// Trigger on HRTIM_TIME compare match for compare register 4
-    TimECmp4 = 26,
-
-    /// Trigger on HRTIM_TIME compare match for compare register 3
-    TimECmp3 = 25,
-
-    /// Trigger on HRTIM_TIME compare match for compare register 2
-    TimECmp2 = 24,
-
-    // /// Trigger on D reset and counter roll-over
-    // _ = 23
-    /// Trigger on HRTIM_TIMD period
-    TimDPeriod = 22,
-
-    /// Trigger on HRTIM_TIMD compare match for compare register 4
-    TimDCmp4 = 21,
-
-    /// Trigger on HRTIM_TIMD compare match for compare register 2
-    TimDCmp2 = 20,
-
-    // /// Trigger on D reset and counter roll-over
-    // _ = 19
-    /// Trigger on HRTIM_TIMC period
-    TimCPeriod = 18,
-
-    /// Trigger on HRTIM_TIMC compare match for compare register 4
-    TimCCmp4 = 17,
-
-    /// Trigger on HRTIM_TIMC compare match for compare register 2
-    TimCCmp2 = 16,
-
-    /// Trigger on HRTIM_TIMB period
-    TimBPeriod = 15,
-
-    /// Trigger on HRTIM_TIMB compare match for compare register 4
-    TimBCmp4 = 14,
-
-    /// Trigger on HRTIM_TIMB compare match for compare register 2
-    TimBCmp2 = 13,
-
-    /// Trigger on HRTIM_TIMA period
-    TimAPeriod = 12,
-
-    /// Trigger on HRTIM_TIMA compare match for compare register 4
-    TimACmp4 = 11,
-
-    /// Trigger on HRTIM_TIMA compare match for compare register 2
-    TimACmp2 = 10,
-
-    // /// Trigger on EEV10
-    // _ = 9,
-
-    // ///  Trigger on EEV9
-    // _ = 8,
-
-    // ///  Trigger on EEV8
-    // _ = 7,
-
-    // ///  Trigger on EEV7
-    // _ = 6,
-
-    // ///  Trigger on EEV6
-    // _ = 5,
-    /// Trigger on HRTIM_MASTER period
-    MasterPeriod = 4,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 4
-    MasterCmp4 = 3,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 3
-    MasterCmp3 = 2,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 2
-    MasterCmp2 = 1,
-
-    /// Trigger on HRTIM_MASTER compare match for compare register 1
-    MasterCmp1 = 0,
-}
-
-macro_rules! impl_flt_monitor {
-    ($($t:ident: ($fltx:ident, $fltxc:ident),)+) => {$(
-        pub struct $t {
-            _x: PhantomData<()>
-        }
-
-        impl FaultMonitor for $t {
-            fn is_fault_active(&self) -> bool {
-                let common = unsafe { &*HRTIM_COMMON::ptr() };
-                common.isr.read().$fltx().bit()
-            }
-
-            fn clear_fault(&mut self) {
-                let common = unsafe { &*HRTIM_COMMON::ptr() };
-                common.icr.write(|w| w.$fltxc().set_bit());
-            }
-
-            // TODO: Should we have our own trait since it does not seem possible to implement this
-            fn set_fault(&mut self) {
-                todo!()
-            }
-        }
-    )+};
-}
-
-impl_flt_monitor!(
-    FltMonitorSys: (sysflt, sysfltc),
-    FltMonitor1: (flt1, flt1c),
-    FltMonitor2: (flt2, flt2c),
-    FltMonitor3: (flt3, flt3c),
-    FltMonitor4: (flt4, flt4c),
-    FltMonitor5: (flt5, flt5c),
-    FltMonitor6: (flt6, flt6c),
-);
